+++ CMakeLists.txt
@@ -2,7 +2,11 @@ cmake_minimum_required(VERSION 3.17)
 
 project(BlindSpot)
 
-add_subdirectory(src)
+set(CMAKE_CXX_STANDARD 23)
+ADD_SUBDIRECTORY(src)
+INCLUDE_DIRECTORIES(include)
+
+
 
 #[===[
 file(GLOB SOURCES
