+++ src/LidarEngine.cpp
@@ -1,5 +1,6 @@
+#include <OpenGL/gl3.h>
 #include <glad/glad.h>
-#include <GL/glut.h>
+#include <GL/freeglut.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/constants.hpp>
